+++ tests/testthat/test-calc_rebound_data.R
@@ -96,7 +96,7 @@ test_that("calc_hat(use_sub_approx = TRUE) works as expected", {
     calc_hat(use_sub_approx = TRUE)
 
   expect_equal(res[[ReboundTools::hat_vars$eta_engr_units_hat]][[1]], 42)
-  expect_equal(res[[ReboundTools::hat_vars$eta_engr_units_hat]][[2]], 81800)
+  expect_equal(res[[ReboundTools::hat_vars$eta_engr_units_hat]][[2]], 81.8)
     
   expect_equal(res[[ReboundTools::hat_vars$eta_hat]][[1]], 0.33169688306808242650)
   expect_equal(res[[ReboundTools::hat_vars$eta_hat]][[2]], 22722.222222222)
@@ -143,7 +143,7 @@ test_that("calc_hat(use_sub_approx = FALSE) works as expected", {
     calc_hat()
   
   expect_equal(res[[ReboundTools::hat_vars$eta_engr_units_hat]][[1]], 42)
-  expect_equal(res[[ReboundTools::hat_vars$eta_engr_units_hat]][[2]], 81800)
+  expect_equal(res[[ReboundTools::hat_vars$eta_engr_units_hat]][[2]], 81.8)
   
   expect_equal(res[[ReboundTools::hat_vars$eta_hat]][[1]], 0.33169688306808242650)
   expect_equal(res[[ReboundTools::hat_vars$eta_hat]][[2]], 22722.222222222)
@@ -191,7 +191,7 @@ test_that("calc_bar() works as expected with approximated hat", {
     calc_bar()
   
   expect_equal(res[[ReboundTools::bar_vars$eta_engr_units_bar]][[1]], 42)
-  expect_equal(res[[ReboundTools::bar_vars$eta_engr_units_bar]][[2]], 81800)
+  expect_equal(res[[ReboundTools::bar_vars$eta_engr_units_bar]][[2]], 81.8)
 
   expect_equal(res[[ReboundTools::bar_vars$eta_bar]][[1]], 0.33169688306808242650)
   expect_equal(res[[ReboundTools::bar_vars$eta_bar]][[2]], 22722.222222222)
@@ -237,7 +237,7 @@ test_that("calc_tilde() works as expected with approximated hat", {
     calc_tilde()
 
   expect_equal(res[[ReboundTools::tilde_vars$eta_engr_units_tilde]][[1]], 42)
-  expect_equal(res[[ReboundTools::tilde_vars$eta_engr_units_tilde]][[2]], 81800)
+  expect_equal(res[[ReboundTools::tilde_vars$eta_engr_units_tilde]][[2]], 81.8)
   
   expect_equal(res[[ReboundTools::tilde_vars$eta_tilde]][[1]], 0.33169688306808242650)
   expect_equal(res[[ReboundTools::tilde_vars$eta_tilde]][[2]], 22722.222222222)