+++ tests/testthat/test-graphs.R
@@ -1,22 +1,22 @@
 
-test_that("rebound_graphs() works as expected", {
+test_that("path_graphs() works as expected", {
   # This is a mess, because all graphs are on the same plot.
   graphs <- load_eeu_data() %>% 
     rebound_analysis() %>% 
-    rebound_graphs()
+    path_graphs()
   expect_true(!is.null(graphs))
   
   # Try with only one type of graph
   graphs_energy <- load_eeu_data() %>% 
     rebound_analysis() %>% 
-    rebound_graphs(graph_types = "Energy")
+    path_graphs(graph_types = "Energy")
   expect_true(!is.null(graphs_energy))
   expect_equal(graphs_energy$plot_env$.path_data$graph_type %>% unique() %>% as.character(), "Energy")
   
   # Try with only one case, Car Energy
   graphs_car_energy <- load_eeu_data() %>% 
     rebound_analysis() %>% 
-    rebound_graphs(cases = "Car", 
+    path_graphs(cases = "Car", 
                    graph_types = "Energy")
   expect_true(!is.null(graphs_car_energy))
   expect_equal(graphs_car_energy$plot_env$.path_data$Case %>% unique(), "Car")
@@ -25,7 +25,7 @@ test_that("rebound_graphs() works as expected", {
   # Try Car Cost
   graphs_car_cost <- load_eeu_data() %>% 
     rebound_analysis() %>% 
-    rebound_graphs(cases = "Car", 
+    path_graphs(cases = "Car", 
                    graph_types = "Cost")
   expect_true(!is.null(graphs_car_cost))
   expect_equal(graphs_car_cost$plot_env$.path_data$Case %>% unique(), "Car")
@@ -34,7 +34,7 @@ test_that("rebound_graphs() works as expected", {
   # Try indexed Car Cost
   graphs_car_cost <- load_eeu_data() %>% 
     rebound_analysis() %>% 
-    rebound_graphs(indexed = TRUE,
+    path_graphs(indexed = TRUE,
                    cases = "Car", 
                    graph_types = "Cost")
   expect_true(!is.null(graphs_car_cost))
@@ -44,7 +44,7 @@ test_that("rebound_graphs() works as expected", {
   # Eliminate the grids for Car Cost graph.
   graphs_car_cost_no_grids <- load_eeu_data() %>% 
     rebound_analysis() %>% 
-    rebound_graphs(cases = "Car", 
+    path_graphs(cases = "Car", 
                    graph_types = "Cost", 
                    grid_types = NULL)
   expect_true(!is.null(graphs_car_cost_no_grids))
@@ -53,7 +53,7 @@ test_that("rebound_graphs() works as expected", {
   # Try an Energy graph for lamps
   graphs_lamp_energy <- load_eeu_data() %>% 
     rebound_analysis() %>% 
-    rebound_graphs(cases = "Lamp", 
+    path_graphs(cases = "Lamp", 
                    graph_types = "Energy")
   expect_true(!is.null(graphs_lamp_energy))
   expect_equal(graphs_lamp_energy$plot_env$.path_data$Case %>% unique(), "Lamp")
@@ -62,7 +62,7 @@ test_that("rebound_graphs() works as expected", {
   # Try an indexed Energy graph for lamps
   graphs_indexed_lamp_energy <- load_eeu_data() %>% 
     rebound_analysis() %>% 
-    rebound_graphs(indexed = TRUE, 
+    path_graphs(indexed = TRUE, 
                    cases = "Lamp", 
                    graph_types = "Energy")
   expect_true(!is.null(graphs_indexed_lamp_energy))
@@ -72,7 +72,7 @@ test_that("rebound_graphs() works as expected", {
   # Try two cases
   graphs_two_cases_indexed_energy <- load_eeu_data() %>% 
     rebound_analysis() %>% 
-    rebound_graphs(indexed = TRUE,
+    path_graphs(indexed = TRUE,
                    graph_types = "Energy") +
     ggplot2::facet_wrap(facets = "Case")
   expect_true(!is.null(graphs_two_cases_indexed_energy))
@@ -80,7 +80,7 @@ test_that("rebound_graphs() works as expected", {
   # Request both cases
   graphs_two_cases_indexed_energy_2 <- load_eeu_data() %>% 
     rebound_analysis() %>% 
-    rebound_graphs(cases = c("Car", "Lamp"), 
+    path_graphs(cases = c("Car", "Lamp"), 
                    indexed = TRUE,
                    graph_types = "Energy") +
     ggplot2::facet_wrap(facets = "Case")
@@ -89,16 +89,16 @@ test_that("rebound_graphs() works as expected", {
   # Try a preferences graph for lamps
   graphs_lamp_prefs <- load_eeu_data() %>% 
     rebound_analysis() %>% 
-    rebound_graphs(cases = "Lamp", 
+    path_graphs(cases = "Lamp", 
                    graph_types = "Preferences")
   expect_true(!is.null(graphs_lamp_prefs))
   
   # Try a preferences graph for lamps with fewer indifference curve points
-  graph_prefs <- ReboundTools::default_graph_params
+  graph_prefs <- ReboundTools::path_graph_params
   graph_prefs$n_indiff_curve_points <- 200
   graphs_lamp_prefs_2 <- load_eeu_data() %>% 
     rebound_analysis() %>% 
-    rebound_graphs(cases = "Lamp", 
+    path_graphs(cases = "Lamp", 
                    graph_types = "Preferences", graph_params = graph_prefs) +  
     ggplot2::xlim(0.9, 2.5)
     ggplot2::ylim(0.99, 1.003)
@@ -281,12 +281,12 @@ test_that("rebound_graphs_helper() works with a preferences graph with grids for
 
 
 test_that("graphs work without arrows", {
-  no_arrows <- ReboundTools::default_graph_params
+  no_arrows <- ReboundTools::path_graph_params
   no_arrows$show_arrows <- FALSE
   # Try with only one case, Car Energy
   graphs_car_energy <- load_eeu_data() %>% 
     rebound_analysis() %>% 
-    rebound_graphs(cases = "Car", 
+    path_graphs(cases = "Car", 
                    graph_types = "Energy", 
                    graph_params = no_arrows)
   expect_true(!is.null(graphs_car_energy))
@@ -296,13 +296,88 @@ test_that("graphs work without arrows", {
 
 
 test_that("points_atop_paths works as expected", {
-  points_beneath_paths <- ReboundTools::default_graph_params
+  points_beneath_paths <- ReboundTools::path_graph_params
   points_beneath_paths$points_atop_paths <- FALSE
   # Try with only one case, Car Energy
   graphs_car_energy <- load_eeu_data() %>% 
     rebound_analysis() %>% 
-    rebound_graphs(cases = "Car", 
+    path_graphs(cases = "Car", 
                    graph_types = "Energy", 
                    graph_params = points_beneath_paths)
   expect_true(!is.null(graphs_car_energy))
 })
+
+
+test_that("sensitivity_graphs() works as expected", {
+  orig_data <- load_eeu_data()
+  sens_params <- list(Car = list(k = seq(0.5, 1.5, by = 0.5)),
+                      Lamp = list(k = seq(0, 2, by = 1)))
+  g <- sensitivity_graphs(rebound_data = orig_data, parameterization = sens_params,
+                          x_var = "k", y_var = "Re_tot") +
+    ggplot2::facet_wrap(facets = "Case", scales = "free_x") +
+    ggplot2::scale_colour_manual(values = c(Re_tot = "black"), guide = FALSE) + 
+    ggplot2::scale_size_manual(values = c(Re_tot = 0.5), guide = FALSE) + 
+    ggplot2::scale_linetype_manual(values = c(Re_tot = "solid"), guide = FALSE) +
+    ggplot2::labs(colour = ggplot2::element_blank(),
+                  size = ggplot2::element_blank(),
+                  linetype = ggplot2::element_blank())
+  expect_true(!is.null(g))
+})
+
+
+test_that("sensitivity_graphs() works with more than 1 line variation", {
+  orig_data <- load_eeu_data()
+  sens_params <- list(Car = list(k = seq(0, 2, by = 0.5), 
+                                   I_E = seq(2, 5, by = 1), 
+                                   e_qs_ps_UC = seq(-0.5, -0.1, by = 0.1)), 
+                        Lamp = list(k = seq(0, 2, by = 0.5),
+                                    I_E = seq(2, 5, by = 1), 
+                                    e_qs_ps_UC = seq(-0.5, -0.1, by = 0.1)))
+  g <- sensitivity_graphs(rebound_data = orig_data, parameterization = sens_params, 
+                     x_var = "I_E", y_var = "Re_tot", line_var = "Case") +
+    ggplot2::facet_grid(rows = ggplot2::vars(k), 
+                        cols = ggplot2::vars(e_qs_ps_UC), scales = "free_y") +
+    ggplot2::scale_colour_manual(values = c(Car = "black", Lamp = "red")) + 
+    ggplot2::scale_size_manual(values = c(Car = 0.5, Lamp = 1.0)) + 
+    ggplot2::scale_linetype_manual(values = c(Car = "solid", Lamp = "dashed")) + 
+    ggplot2::labs(colour = ggplot2::element_blank(), 
+                  size = ggplot2::element_blank(),
+                  linetype = ggplot2::element_blank())
+  expect_true(!is.null(g))
+})
+
+
+test_that("rebound_terms_graph() works as expected", {
+  df <- load_eeu_data()
+  sens_params <- list(Car = list(eta_engr_units_star = seq(35, 50, by = 0.5)), 
+                      Lamp = list(eta_engr_units_star = seq(70, 90, by = 5)))
+  g <- rebound_terms_graph(rebound_data = df, parameterization = sens_params, 
+                      x_var = "eta_engr_units_tilde") +
+    ggplot2::facet_wrap(facets = "Case", scales = "free_x")
+  expect_true(!is.null(g))
+  g2 <- rebound_terms_graph(rebound_data = df, parameterization = sens_params, 
+                           x_var = "eta_engr_units_tilde", include_Re_tot = FALSE) +
+    ggplot2::facet_wrap(facets = "Case", scales = "free_x")
+  expect_true(!is.null(g2))
+})
+
+
+test_that("sensitivity graphs correctly order points", {
+  df <- load_eeu_data()
+  eta_sens_params = list(Car = list(eta_engr_units_star = seq(35, 50, by = 0.5)), 
+                         Lamp = list(eta_engr_units_star = seq(70, 90, by = 5)))
+  
+  g <- sensitivity_graphs(rebound_data = df, parameterization = eta_sens_params,
+                     x_var = "eta_engr_units_star", y_var = "Re_tot") +
+    ggplot2::facet_wrap(facets = "Case", scales = "free_x") +
+    ggplot2::scale_colour_manual(values = c(Re_tot = "black"), guide = FALSE) + 
+    ggplot2::scale_size_manual(values = c(Re_tot = 0.5), guide = FALSE) + 
+    ggplot2::scale_linetype_manual(values = c(Re_tot = "solid"), guide = FALSE) +
+    ggplot2::labs(x = expression(tilde(eta)*" [mpg (Car) or lm/W (Lamp)]"),
+                  y = expression(Re[tot]*" [-]"),
+                  colour = ggplot2::element_blank(),
+                  size = ggplot2::element_blank(),
+                  linetype = ggplot2::element_blank())
+  expect_true(!is.null(g))
+})
+